Replaced all unicode-printed chars with plain ASCII (Closes: #765886, #665847)
Yes, it is nearly 2016 and Debian has been using unicode for over a
decade. Yes, everybody should be using it by now. What made me (josch)
remove unicode characters and revert to back to ASCII is that the only
advantage of using unicode characters is:

 - Slightly more pretty output. I'd say the improvement is so little
   that I would call it negligible.

But on the other hand, there is a disadvantage which I'd call much more
practical than maybe having a more visually pleasing output:

 - sbuild logs are not necessarily only consumed by users of Debian or
   transmitted with technologies we ship with Debian. They might be
   copy-pasted to mail messages (read by people with very different clients),
   personal book keeping programs and web interfaces of services with no
   unicode support. Yes, all these places *should* know unicode but sbuild
   logs should also not become less readable if they don't.

Given that there is really no hard *requirement* for keeping the unicode
characters for pretty drawing (in contrast to a program which, for
example, handles asian script), I'm applying Dima's patch.

The reason why ASCII output is now the default instead of adding a
--ascii option is, that at the time where the build is done, one does
not know which places the build logs might reach.  The same reasoning
goes for why unicode support is not enabled or disabled depending on the
capabilities of the terminal sbuild outputs to.  Instead, if one is very
sure that the log will never be consumed by or transmitted through a
medium without unicode support, then there could be a --pretty-unicode
option in the future which explicitly enables the pretty boxes again.