external commands are now unparsed strings (Closes: #782672)
when handling external commands (--starting-build-commands and others), I now
keep the input string as is, and pass it to the shell, instead of splitting it
on whitespace and then reassembling it. This reassembling was being done
improperly, with shell internals such as < being erroneously escaped. Keeping
the string as is allows the user to give the command they want.

This is implemented by keeping the COMMAND and INTCOMMAND keys same as they were
(array-refs of strings), but adds additional COMMAND_STR and INTCOMMAND_STR keys
that have the string representation. The non-STR versions are always populated,
and the external command options populate the _STR ones as well. When running
the commands, we use the _STR options if we have them, otherwise, we recombine
as before.

Since the old array-ref-based code path is still there, old commandlines and
configurations remain working