Added --xxx-failed-commands
Added several external commands that trigger on failure. The way these are
intended to be used is to spawn a shell if something goes wrong so that the user
can poke around and find the problem. Most commonly you would do

  sbuild --anything-failed-commands %s

Currently we have

 --build-deps-failed-commands
 --build-failed-commands
 --anything-failed-commands

Where --anything-failed-commands simply activates all of the others