lib/Sbuild/Build.pm: make sure regex only matches lines in the Files field (closes: #808560)
Previously, the regex would also match lines in the Checksums-Sha1 and
Checksums-Sha256 field. This was no problem because prior to Perl 5.22 because
the following list slicing would return an empty array when it found an
out-of-bounds index. Since Perl 5.22 an array of one undef element is returned
instead. More info can be found at:

https://rt.perl.org/Public/Bug/Display.html?id=114498

To fix this we modify the regex such that it only matches the lines containing
a 32 character checksum which should only happen in the lines of the Files
section where the result of calling split() should return an array of
sufficient size to not yield an undef when retrieving element 5 from it.