lib/Sbuild/Chroot.pm: apply environment filter *before* setting environment variables set by sbuild itself (closes: #807488)
Before this commit the ENVIRONMENT_FILTER configuration option was able
to unset environment variables that were set by the sbuild code itself.
This for example undid the setting of the APT_CONFIG environment
variable for apt in ResolverBase.pm. The ENVIRONMENT_FILTER variable
should be to filter the user environment when they call sbuild. It
should still be possible for sbuild itself to enforce an environment
variable to be set when calling run_command() or pipe_command(). If an
environment variable set by sbuild is wrong or undesirable then
ENVIRONMENT_FILTER would be the wrong place to control it.