Do architecture check after build-essential is installed (closes: #700317)
The architecture check is done inside the chroot to not make the
architecture check dependent on a possibly outdated libdpkg-perl on the
host (useful for porters of new architectures). But this means that
libdpkg-perl has to be installed inside the chroot. Thus, the
architecture check can only be done after build-essential has been
installed.