Allow to build by only passing source package name without version (closes: #693928)
The packagename_version format was a handy way to fulfill the assumption
sbuild makes that the source package version is available right from the
start. Sbuild then uses the version for things like log filtering,
status output, external command percentage escapes or the build log file
name.

This patch makes it optional that the version is known from the start.
In case it is not known it will:

 - not print version information in the build log
 - not let external command percentage escapes be defined
 - use only the package name for the build log filename
 - set up log filtering much later when the version is definitely known

To not break existing setups, all the old behavior is kept if the
version is known from the start. To distinguish the version being known
from not being known, the user-supplied argument is checked for the
presence of an underscore. If an underscore exists, then either a dsc is
passed or the pkgname_version format was used and in both cases, the old
behavior can be executed. If no underscore can be found, then it is a
bare package name and the different behavior explained above will kick
in.

The man page has been adapted to document the new behavior.

While I was at it, I also rewrote the code downloading the source
package:

 - The old code triggered an "apt-get update" if the source package
   could not be found. This is not necessary because the source package
   is fetched after the chroot is updated. Furthermore it allows to
   "apt-get update" to be run without eventual failures be caught by the
   respective external command hooks. Lastly, this would ignore the
   --no-apt-update command line option.

 - "apt-cache showsrc" was run without --only-source

 - The "apt-cache showsrc" output was parsed using Perl regexes.
   Dpkg::Index is used now.

 - The Files field is now parsed using Dpkg::Checksums instead of using
   perl regexes.

 - the @fetched array was only filled but never used by anybody