Sbuild::ResolverBase: use copy: for dummy archive
We used to use file: URIs for sbuild's dummy archive, but recent
versions of apt put symlinks in /var/lib/apt/lists/ if we do that, which
means that those symlinks become broken as soon as we remove the dummy
archive.  sbuild itself is OK with this, but launchpad-buildd tries to
read /var/lib/apt/lists/*_Packages in order to do dep-wait analysis
after failures, and that fails if there are broken symlinks there.  It's
easy enough to avoid this problem by using copy: URIs instead, so let's
do that.