bin/sbuild-createchroot: add support for SUITE-VARIANT (closes: #826957)
This is useful to create variants of a base suite like jessie-backports
which which would create a jessie chroot to which the backports archive
can be manually added after the base chroot was created using
sbuild-shell. This way, a normal jessie chroot and a jessie-backports
chroot can exist side-by-side. Besides this common use case, this syntax
can also be used to create chroots with any other type of customization.
It is a shortcut for creating a base chroot and then having to manually
copy it and edit the configuration files.