bin/sbuild-createchroot: add --alias option (closes: #800593)
Add an alternative name that the chroot will be known by. This option can be
given multiple times to add more than one alias. Using an alias is useful for
chroots of distributions that are known by more than one name. For example
Debian unstable is also known as sid. Additionally, sbuild chooses the
distribution by the latest changelog entry which could list UNRELEASED for
packages that the maintainer is currently working on. For Debian it thus makes
sense to add UNRELEASED as an alias for a Debian unstable chroot.