bin/sbuild-createchroot: Add --extra-repository option
Add a repository to the list of apt sources. The repository specification is a
line suitable for an apt sources.list file. For instance, you might use

  --extra-repository="deb http://httpredir.debian.org/debian experimental main"

to allow packages in the experimental distribution to fulfill
build-dependencies. Note that the build chroot must already trust the key of
this repository. This option is especially useful for creating variant chroots.