Use armored ASCII GPG key pairs instead of the GPG binary format and support GPG 2.x inside the chroot
The GPG binary key format is not guaranteed to be compatible between
different versions of GPG. For example the 2.x format is incompatible
with the 1.x format and the other way round. Still, key pairs might be
generated with GPG 1.x outside the chroot and will be used with GPG 2.x
inside the chroot or the other way round. Thus, a solution is needed
which allows GPG 1.x key pairs to be used by GPG 2.x and the other way
round.  One solution is the one implemented by this commit and was
proposed by Daniel Kahn Gillmor in #debian-gnupg (Thanks!!). Instead of
using the binary format, `sbuild-update --keygen` will now export the
key pair in armored ASCII format and GPG inside the chroot will import
the key pair. Since the armored ASCII key format is independent of the
GPG version generating or importing it, it doesn't matter which
combination of GPG is installed on the machine running `sbuild-update
--keygen` or inside the chroot. Thus, no limitations are created when
sbuild is to work on the same host with chroots containing different GPG
versions (for example of the 1.x and 2.x versions).

To implement this new handling of GPG key pairs, the
SBUILD_BUILD_DEPENDS_SECRET_KEY_ARMORED and
SBUILD_BUILD_DEPENDS_PUBLIC_KEY_ARMORED configuration options are
introduced by this commit and they deprecate the
SBUILD_BUILD_DEPENDS_SECRET_KEY and SBUILD_BUILD_DEPENDS_PUBLIC_KEY
option pair.

Sbuild will still be able to work with the old binary format for a
while. It will just not be able to handle chroots with an incompatible
GPG version until `sbuild-update --keygen` has been re-run on the host
to create the armored ASCII keys.

Furthermore, this commit adds support for GPG 2.x inside the chroot.
Those GPG versions will start a gpg-agent process once gpg is run. If
that process is not cleaned up before the chroot is exited, then closing
the chroot might fail because gpg-agent will still have open file
descriptors and thus unmounting filesystems might fail. Thus, if the
gpgconf executable exists, `gpgconf --kill gpg-agent` will be executed
to clean up any remaining gpg-agent processes after operations involving
gpg.