lib/Sbuild/Build.pm/check_space(): Add up correct size when looping over files (closes: #825359)
Due to a possible copy&paste bug, the loop that is supposed to add up
the sizes of the generated build artifacts was instead repeatedly adding
up the size of the unpacked source package, leading to implausible
results as the unpacked source package can be an order of magnitude
bigger in size than each of the generated binary packages. The problem
was pronounced for source packages with many binary packages.