Run apt-cache without --only-source (closes: #834704)
We would like to call apt-cache with --only-source so that the result
only contains source packages with the given name but this feature was
only introduced in apt 1.1~exp10 so it is only available in Debian
Stretch and later. Thus we don't add this flag and instead filter the
generate result to only consider the source packages we are interested
in.