bin/sbuild-createchroot: replace the SUITE-VARIANT syntax with the --chroot-prefix option (closes: #834515)
Introducing the SUITE-VARIANT syntax was a mistake because it makes it
impossible to create a chroot for distribution names that contain the
minus character. So instead of overloading the meaning of the SUITE
argument with the prefix of the resulting chroot name, we add a new
option called --chroot-prefix which works analogous to --chroot-suffix
and defaults to the value of SUITE if not set.