Partly revert 956fc29b and --import old keys instead of trying to use them with --secret-keyring (closes: #827315)
It turns out that gpg2 *can* read the old format, just not with its
--secret-keyring and --keyring arguments but with --import. This makes storing
the keys in armored format as it was introduced by commit 956fc29b useless.
Since key signing is only required for Debian squeeze anyways, we want to avoid
carrying more complexity forward than absolutely needed.

To also cater for chroots without gpg, signing is now also dependent on
gpg being present. No signing is done if the gpg binary doesn't exist.
Since the gnupg package originally was drawn in by apt and only apt
versions of 1.3~exp1 or newer stopped depending on gnupg, it should be
safe to assume that if no gpg is present in the chroot, then also apt is
new enough to understand [trusted=yes].