Stop depending on newlines being produced by the build to detect timeouts
Before this commit, sbuild would read the combined output of
dpkg-buildpackage line-by line. Only when a new line was encountered
would it reset the timeout to kill the build for inactivity. This means,
that processes that produce data on standard output or standard error
but do not include a newline will still get killed after a timeout as no
reset happens despite data being written.

This commit fixes this behaviour. Instead of reading the output of
dpkg-buildpackage line-by-line, we use sysread to acquire data once it
is available. This will reset the timeout every time data is produced by
the build process and not require a newline character to be produced.

As a side-effect, this also allows us to special-case the \r character
as it is used in progress meters for example. If an \r character is
produced, sbuild will now treat it as a line terminator and thus
effectively replace it by a \n in the build log.