Revert 'lib/Sbuild/Build.pm: Fix "W: Stripped single quote from command for security"'
This reverts commit f8efb79c6b70912af19e1e3507e721dd9630606a.

Using double quotes instead of single quotes is no solution to the
problem. The chroot backends really ought to support single quotes as
part of commands to be executed. Thus, the proper fix is, to escape the
command string correctly before executing it. Just disallowing single
quotes is a quick and dirty fix but not the correct one. For more
information and discussion see Debian bugs #816493 and #836175.