Make Chroot ID optional
Chroots other than schroot might not offer a way to enumerate all
available chroots. Think for example about autopkgtest ssh or qemu
backends. This means, that when using the autopkgtest chroot mode, the
user had to manually add --chroot unstable-amd64-sbuild to the
invocation and change that value manually depending on the distribution
and architecture. It would be much more intuitive if the user could
instead specify: --autopkgtest-virt-server-opts="-- %r-%a-sbuild". This
wasn't possible so far, because the manually specified chroot name was
always appended to the autopkgtest options. Doing so has yet another
disadvantage. Some backends like the ssh or qemu backend don't even have
a good definition of a chroot name. It would be very unintuitive to
shoehorn the last command line option given to those backend into the
--chroot parameter. Thus, this commit implements the following changes:

 - the autopkgtest-virt backends are no longer supplied with the chroot
   name
 - computing the Chroot ID from the distribution and architecture has
   been made optional as it is only applicable for backends like schroot