For binNMUs, instead of copying the timestamp of the last changelog entry, generate a new one (closes: #843773)
Earlier implementations used the date of the last changelog entry for
the new entry so that Multi-Arch:same packages would be co-installable
(their shared changelogs had to match). This is not necessary anymore as
binNMU changelogs are now written into architecture specific paths.
Re-using the date of the last changelog entry has the disadvantage that
this will effect SOURCE_DATE_EPOCH which in turn will make the
timestamps of the files in the new package equal to the last version
which can confuse backup programs.  By using the build date for the new
binNMU changelog timestamp we make sure that the timestamps of changed
files inside the new package advanced in comparison to the last version.