Install dose-distcheck and lintian for the native architecture
When crossbuilding, sbuild will try to install the host architecture
versions of dose-distcheck and lintian. This doesn't make much of a
difference for lintian as that package is Architecture:all right now but
dose-distcheck is Architecture:any and the host architecture version
will not be installable and/or executable. Lets explicitly install the
native version instead.