do not output empty color reset tags
In the current configuration, sbuild outputs empty color reset
sequences which might confuse some programs. In particular, it creates
a lot of noise in Emacs where each line is prefixed with a double
"reset" escape sequence, regardless of whether the line is colored or
not.

Here we change the output colors to only be sent if they are not the
"reset" color so that we remove noise from the output.