add --post-build-failed-commands hook
The point of this hook is to make sure there is a way to run a hook
outside the chroot in the case the build succeeds (which
post-build-commands does not) or fails (which it does not).

My main use case for this is to send user-visible notifications (with,
for example. notify-send) when a build completes, regardless of
state.