Add configuration option to keep downloaded packages by APT (Closes: #933723)
sbuild was always setting APT::Keep-Downloaded-Packages to false which caused
downloaded packages to be always deleted from the APT cache after installation.
This prevented to take advantage of using a persistent package cache to avoid
downloading packages on every sbuild invocation.

The added option "apt_keep_downloaded_packages" is 0 by default, which cause
sbuild to have the same behavior as before, that is setting
APT::Keep-Downloaded-Packages to false.