verify() in unshare uses the same (wrong) parameter for $verify_*
Furthermore, the "+" expansion expands to "" if the input is "", but the
alternate ("no") otherwise, so the variables will *never* be "yes"
(making the later tests against the value "yes" contradictions).

Rather than using a bunch of boolean yes/no arguments, I propose naming
the pieces expected to be verified.

Further, the .dsc files shouldn't be verified when sbuild *without*
--source is called as, though the .dsc file is produced, it is not
signed in this circumstances (as it's absent from the .changes file).