Update verify_deb to match current zstd output
In 1.5.2, zstd changed its output to produces different headers (and
footers). Then, in 1.5.4 (current in unstable at the time of writing),
zstd changed again such that the compressed data itself changed.

We could unpack things for verification like verify_orig does, but while
the compression of the orig tarball changing doesn't really affect the
reproducibility of builds, the compression of the deb changing almost
certainly does. Hence I suspect the only option here is to update the
test to match the current output of zstd and hope zstd doesn't change
too often in the future.